pyproject.toml
[project]
name = "thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lightrag-hku>=1.4.8.2",
    "plastex>=3.1",
    "tensorflow",
    "numpy",
    "pandas",
    "scipy",
    "matplotlib",
    "pyyaml",
    "scikit-learn",
    "scikit-image",
    "pathlib2",
    # MDP Topology Optimizer dependencies
    "deap",  # Distributed Evolutionary Algorithms in Python
    "seaborn",  # Enhanced visualization
    "tqdm",  # Progress bars
    "imageio",  # Image processing
    "Pillow",  # Image processing
    # OpenRouter and embeddings dependencies
    "openai",  # OpenRouter API client
    "sentence-transformers",  # Fast embeddings
    "python-dotenv",  # Environment variables
    "tf-keras",  # Keras compatibility for transformers
]
